WIP: Restructure routing to allow top level routes
The problem was to have a login page without doing magics to hide the
user-bar and nav-bar the routing had to be restructured such that
you can place a component at /login and only that component's template
would be displayed.

A HomeModule has been added to group the components and modules
that can be used when authenticated.

A LoginComponent was added for displaying and managing the Login Page.

A VolumesComponent was added because I thought I could simply
import the VolumesModule and it would be a child of HomeComponent
in terms of routing. That is, without knowledge of the sub routes
defined in volumes.routing-module.ts one could access /volumes/list
and it would display in the <router-outlet> defined in
HomeComponent. This is however not the case. The child routes
need to be known and registered in home-routing.module.

The commit enables many debug output and removes lazy loading since
it was crashing without a clear error message.