Refresh the current route if null address was cached
It may occur that the Docks API address is not in local
storage, causing components to make requests using a null
value. When the address is fetched from the server we
refresh the current route to force components to make
initialisation requets again.

TokenStorage has been replaced with StorageService.
This class is identical to TokenStorage and was done for
debugging. For some reason TokenStorage/StorageService
cannot be provided in the tests.